+++ R/convert_CW2pLM.R
@@ -27,7 +27,7 @@
 #' The transformation is recommended for curves recorded with a channel
 #' resolution of at least 0.05 s/channel.
 #'
-#' @section Function version: 0.4.2
+#' @section Function version: 0.4.3
 #'
 #' @author
 #' Sebastian Kreutzer, F2.1 Geophysical Parametrisation/Regionalisation, LIAG - Institute for Applied Geophysics (Germany)
@@ -69,30 +69,37 @@
 #'
 #' @export
 convert_CW2pLM <- function(
-  values
+  object,
+  ...
 ) {
   .set_function_name("convert_CW2pLM")
   on.exit(.unset_function_name(), add = TRUE)
 
+  ## deprecated argument
+  if ("values" %in% ...names()) {
+    object <- list(...)$values
+    .deprecated(old = "values", new = "object", since = "1.2.0")
+  }
+
   ## Integrity checks -------------------------------------------------------
 
   ##(1) data.frame or RLum.Data.Curve object?
-  .validate_class(values, c("data.frame", "RLum.Data.Curve"))
-  .validate_not_empty(values)
-  if (ncol(values) < 2) {
-    .throw_error("'values' should have 2 columns")
+  .validate_class(object, c("data.frame", "RLum.Data.Curve"))
+  .validate_not_empty(object)
+  if (ncol(object) < 2) {
+    .throw_error("'object' should have 2 columns")
   }
 
   ##(2) if the input object is an 'RLum.Data.Curve' object check for allowed curves
-  if (inherits(values, "RLum.Data.Curve")) {
-    if(!grepl("OSL", values@recordType) & !grepl("IRSL", values@recordType)){
-      .throw_error("recordType ", values@recordType,
+  if (inherits(object, "RLum.Data.Curve")) {
+    if(!grepl("OSL", object@recordType) & !grepl("IRSL", object@recordType)){
+      .throw_error("recordType ", object@recordType,
                    " is not allowed for the transformation")
     }
-    temp.values <- as(values, "data.frame")
+    temp.values <- as(object, "data.frame")
 
   }else{
-    temp.values <- values
+    temp.values <- object
   }
 
   # Calculation -------------------------------------------------------------
@@ -109,13 +116,13 @@ convert_CW2pLM <- function(
   # Return values -----------------------------------------------------------
 
   ##returns the same data type as the input
-  if (is.data.frame(values)) {
+  if (is.data.frame(object)) {
     return(temp.values)
   }
 
   set_RLum(
       class = "RLum.Data.Curve",
-      recordType = values@recordType,
+      recordType = object@recordType,
       data = as.matrix(temp.values),
-      info = values@info)
+      info = object@info)
 }